refactor: 명확한 표현을 위해 `Line`에서 `RowLine`으로 변경
